Set and clear thd local 'should_update_hlc' correctly.
Summary:
Enabling and disabling of HLC is controlled by a system variable
'enable_binlog_hlc'. To prevent ongoing trxs from corrupting binlogs, this
variable is captured in a thread local when the first event of the trx is
written to the trx binlog cache and cleared during the commit time of the trx.
However, the thd local was not cleared during the trx 'rollback' path. This
causes problems when the same thread gets used for a subsequent trx - the begin
of a trx will not write a dummy metadata event in the cache, but the commit of
the trx (which relies on thread local to figure out if hlc is enabled) writes
metadata event in the trx cache (effectively corrupting one trx in the binlog).

This fix correctly unsets the thread local when we write the first event for the
trx.

Reference Patch: facebook/mysql-5.6@8cd3411

Originally Reviewed By: abhinav04sharma

fbshipit-source-id: f070e22