Bypass ACL DDLs in gap lock error check [RocksDB part]
Summary:
In MySQL 8.0, auto commit is turned off for both ACL DDL and
DDL queries to avoid unnecessary commits in DD. For DDLs, auto
commit can be turned back on in DD, but not for ACL queries if
they don't go into DD. As a result, the ACL query is treated as
multi-statement query when it comes to checking the gap lock hence
is prohibited. This commit bypasses ACL queries when checking the
gap lock.

Differential Revision: D18920694