Fix jemalloc linking for gunit tests
Summary:
Update the cmake files to use the same link flags for mysqld as for the
gunit tests, so that if jemalloc is linked in, the gunit tests will also
get access to them. The symbol malloc_stats_print is in sql_parse, which
is part of libsql_main.

Squash with D7337802

Reviewed By: lth

Differential Revision: D19169880

fbshipit-source-id: 1df23dad4a2