Adding 2 variables for apply log and apply index
Summary:
Initially we made apply log names derivable
from binlog name, because we did not want the burden of
adding these 2 variables. However this causes problems for MTR
where the rules of naming are different.

Adding 2 variables to make this configurable. In the future
we will make sure the rule checks are only for Raft mode.

Reviewed By: abhinav04sharma

Differential Revision: D19226962

fbshipit-source-id: ad93026f821