Fix valgrind errors
Summary:
We ported 5.6 valgrind suppression for RocksDB ThreadStatusUtils leaks to 8.0.17 but unfortunately it didn't work - the ported suppressions are old suppressions that refer to old RocksDB code and no longer work. And more confusingly, 5.6 has more suppressions in valgrind.supp for the exact same leaks later (since the earlier suppressions stopped working). As a result, 5.6 has duplicated set of suppressions for the same leaks, and we ported the earlier ones that didn't work.

This updates the suppression to work with latest RocksDB code and I choose to not to port the duplicates. We can choose to clean up the duplicated suppressions in 5.6 as well but that seems unnecessary at this point.

We can also look into fixing the leak in RocksDB later - at least the leak in DBImpl::Open seems to be straight-forward. Will file a task to RocksDB later.

Squash with: D19216645

Reviewed By: luqun

Differential Revision: D19247397

fbshipit-source-id: 6e81e20b781