Adding a function in replication.h to sync update GTID sets in mi->rli
Summary:
This is the equivalent of add_logged_gtids in queue_event.
Before a trx is appended to relay log, the logged gtid sets is updated
with the gtid. This ensures that SHOW SLAVE STATUS refledcts up to
date info, but more importantly keeps the previous gtid set correct
on the next file during rotation.

Reviewed By: abhinav04sharma

Differential Revision: D19252866

fbshipit-source-id: e4e7ce79a1c