Fix main.ipv4_as_ipv6
Summary:
Fixes https://bugs.mysql.com/bug.php?id=98177

After connection, we weren't calling vio_io_wait and checking for errors afterwards before starting to write to network.

For some reason, this code was removed when the async client was upstreamed. The fix is just to add it back in.

Reviewed By: zhichengzhu

Differential Revision: D19341643

fbshipit-source-id: 51aa0f13121