Rotate events on slaves cannot go through consensus
Summary:
When we are in ordered commit or rotation functions of
a mysql slave, rotate_using_raft should be false, because we
don't want to run rounds of consensus from the follower.

We add a boolean is_apply_log = false. This is the more correct
way of checking whether something is a raft apply log, rather
than enable_raft_plugin && !thd->rli_slave.

Reviewed By: abhinav04sharma

Differential Revision: D19355147

fbshipit-source-id: f31a984915a