Fixing issue of plugin staying in INACTIVE state
Summary:
In a previous diff we  postponed the initialization
of raft plugin init to a last step post recovery. This is for
DMP/restarts so that engine recovery and binlog recovery is
already done. We however made the mistake to gate the plugin
initialization with enable_raft_plugin. Plugins have to be
initialized completely, otherwise they remain in "INACTIVE"
state and then you can neither uninstall them nor complete
the initialization process.

Reviewed By: abhinav04sharma

Differential Revision: D19420938

fbshipit-source-id: 623eba47dcb