Return session tracker information in same query result
Summary:
this is a port of D18866865

Previously we only returned session state changes
on the final statement of a query. In the multiquery case,
if an intermediate query errors out, we don't get to see
the session change flag. Send the flag with the same
statement that first triggered the change to avoid this case

Since the original function we piggy backed onto was changed,
add a new function and call it in the same paths for COM_QUERY
This does not add it to the stored_procedure paths

Upstream is tracking a request to add this to the upstream protocol at https://bugs.mysql.com/bug.php?id=97885

Squash with: D18851147

Reviewed By: jkedgar

Differential Revision: D19626226

fbshipit-source-id: 6e5a3ef8a38