Switch to static linking libssl_pic.a for BoringSSL
Summary:
BoringSSL are statically linked with libmysqclient/libmysqlrouterl/etc and therefore requires _PIC version - and we don't have boringssl.so so we can't dynamic link.

Without this change we are dynamically linking libssl.so.

Interestingly, note that with "Add PIC libraries for OpenSSL" change (that was reverted later), given that we are probing against ssl_pic, there is no ssl_pic.so or libssl_pic.so, so cmake end up choosing libssl_pic.a which means static linking OpenSSL and you can end up with multiple copies. However, interestingly for OpenSSL all the statically linked symbols are public visible (T), so they end up causing symbol conflicts. Case in point, Herman was seeing that router_component_mock_server is failing because module A is calling into OPENSSL_cleanup in A which was resolved to another module B's OPENSSL_cleanup copy, and that function calls atexit(ssl_library_stop function) which would means B!ssl_library_stop would get run at shutdown, but when that gets run B is already unloaded causing a crash.

For BoringSSL, it is somewhat better than OpenSSL because BoringSSL is built with -fvisibility=hidden, so all statically linked symbols are local (t). This means you won't get symbol conflict, and if every module only work with its own copy, no bad things will happen. However, once modules start passing around pointers, all bets are off. I'm seeing routertest_component_http_server test failing because it tries to connect against default cipher in one of the scenarios, and BoringSSL failed to find the cipher in the cipher list (though it is actually there in the list) because there are two copies of BoringSSL. The cipher list from SSL_CTX is allocated in libmysqlrouter_http.so while the cipher list that was comparing against is libhttpclient.so, and it's done with basic pointer comparisons, which naturally fails with both have its own copy of the cipher list bssl::kCiphers. However given that this is limited to the unit test scenario, and we are doing static linking in 5.6 today as well, this should be acceptable, and we can disable the failing scenarios that we don't care about.

Squash with D17216251

Reviewed By: Pushapgl

Differential Revision: D19385368

fbshipit-source-id: 156fb32e3de