Better column family option parsing error in rocksdb
Summary:
Today if you have a super long column family option in rocksdb_default_cf_options / rocksdb_override_cf_options / rocksdb_update_cf_options that has an invalid argument, it's hard to know which option is the culrprit as the error message in MyRocks error/warning log doesn't tell you the invalid option though RocksDB does. This adds the correct error from RocksDB.

It would also be nice if the error would show up directly as mysql error instead of in mysql error/warning log when setting rocksdb_update_cf_option (the other two are read-only), but that is a task for another day.

This is back ported from 8.0 as I've seen simlar issues happening for 5.6 so I thought it'd be a good idea to include it here as well.

Reviewed By: zhichengzhu

Differential Revision: D19635336

fbshipit-source-id: b24cfd1f421