FB8-242: Force minimal required version of liblz4 (1.8.0) and libzstd (1.4.0) (#1086)
Summary:
https://github.com/facebook/mysql-5.6/commit/e9f1b50940b introduced changes that require newer versions of liblz4 (1.8.0) and libzstd (1.4.0).
This causes compilation errors with older libs e.g. for Ubuntu Bionic which has liblz4-dev 1.7.1 and libzstd-dev 1.3.3.
More information at https://jira.percona.com/browse/FB8-242

This patch solves the issues with the following changes:
1. Check minimal required version of liblz4 (1.8.0) and libzstd (1.4.0)
2. Update bundled liblz4 to 1.8.0
3. Add bundled libzstd 1.4.0

Notes:

Some of this can be dropped when moving to 8.0.19+, where the bundled
versions of zstd have been updated to 1.4.0. Manuel mentioned the
changes could also be made to be compatible with older versions of
the library, so the changes can be made for the next rebase.

Squash with D19316704?

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1086

Reviewed By: lth

Differential Revision: D19398515

Pulled By: lth

fbshipit-source-id: 72f063a7a30