Adding updateConsensus hook to binlog rotation using raft
Summary:
This addresses a nasty bug that was observed during
myshadow testing. When we rotate files, the rotate event is
never reported to the plugin via "updateConsensus". This causes
a gap in the applied marker, which causes a deadlock/hang when
this instance is asked to become the master again. The hang
happens in SwitchMySQLToMaster

Reviewed By: abhinav04sharma

Differential Revision: D19484714

fbshipit-source-id: 7da3f38ec45