Use performance_schema.variables_by_thread for rpl_slave_tx_isolation
Summary:
This updates rpl_slave_tx_isolation test to use performance_schema.variables_by_thread and avoid printing out global tx_isolation as the original test was intended to test the transaction_isolation (used to be tx_isolation in 5.6) in the slave threads.

Unfortunately this still doesn't work in 8.0 today as performance_schema.variables_by_thread doesn't show variables from slave SQL/IO threads, so this still isn't complete. But at least this aligns with the direction we want to go and doesn't accidentally impact other changes (such as when we introduce tx_isolation as compat for 5.6 java clients and this test will accidentally fail).

Squash with D19036547

Reviewed By: lth

Differential Revision: D19586194

fbshipit-source-id: a44eabbd752