Fix replication type conversion tests timeout by reverting default of rbr type mismatch whitelist to empty
Summary:
A lot of type conversion related tests are failing because we flipped the RBR type mismatch whitelist to .* but many tests are expecting the conversion to fail and as a result stuck in waiting for the failure, hence the timeout.
The fix is to revert the default back to empty.
Another option is to change the tests / suite options and only set it to empty in those cases, but it makes more sense to me for the defaults to map to 'official' MySQL 8.0 behavior and change to FB specific value in prod instead.

Squash with D19292075

Reviewed By: abhinav04sharma

Differential Revision: D19613856

fbshipit-source-id: 52302e937d4