fix main.events_1
Summary:
as title said

We can't alter a event happened in the past
mysqltest: At line 69: Query 'ALTER EVENT event_starts_test ON SCHEDULE AT '2020-02-02 20:00:02'' failed.
ERROR 1589 (HY000): Event execution time is in the past and ON COMPLETION NOT PRESERVE is set. The event was not changed. Specify a time in the future.

Reviewed By: yizhang82

Differential Revision: D19727591

fbshipit-source-id: 5f9550637a3