Fix Regex_list_handler
Summary:
When the Regex_list_handler was added in 5.6, the key (first parameter) to mysql_rwlock_init() was included for HAVE_PSI_INTERFACE but was excluded when this compile flag was not set.  This required the Regex_list_handler to be done different ways for the two different modes.  In 8.0 the key is always supplied, so we can simplify the Regex_list_handler code and get rid of some compile errors when HAVE_PSI_INTERFACE is disabled.

Squash with: D61551

Reviewed By: lth

Differential Revision: D19821914

fbshipit-source-id: 2bc4ca9c166