Issue #108: Index-only scans do not work for partitioned tables and extended keys
Summary:
- Introduce handler::init_with_fields() - a type of init function
  in the storage engine that is called *after* TABLE_SHARE's field
  and key objects are populated.

- Add an implementation in ha_rocksdb which sets
  HA_PRIMARY_KEY_IN_READ_INDEX when PK datatypes allow this

- Add an implementation in ha_partition which invokes init_with_fields
  for underlying partitions

- Adjust table.cc:open_binary_frm() so that it calls h->init_with_fields()
  and does a relevant part of "Extended keys" processing after that.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/070a257a1c3

Squash with: D51195

-----
Porting Notes

It's not entirely clear if many of the changes here are necessary. But I'm porting the change anyway because MyRocks code is already on this patch (see init_with_fields, since I ported the MyRocks portion separately) so this is the MySQL portion of the port which is needed to align with MyRocks changes.

A bit of history...

In https://github.com/facebook/mysql-5.6/commit/06f9e0c sql/table.cc we've added following line to refresh the value of ha_option given that we may include HA_PRIMARY_KEY_IN_READ_INDEX:

```
ha_option= handler_file->ha_table_flags();
```

So that this line will get a chance to execute:

```
          if (handler_file->table_flags() & HA_PRIMARY_KEY_IN_READ_INDEX) {
            if (field->key_length() == key_part->length &&
                !(field->flags & BLOB_FLAG))
              field->part_of_key = share->keys_in_use;
            if (field->part_of_sortkey.is_set(key))
              field->part_of_sortkey = share->keys_in_use;
          }
```

However, it's not clear if this is required in 8.0 today or even back in 5.6. Reading the code in `prepare_share` in dd_table_share.cc and `open_binary_frm` in table.cc, just 40 or so lines down we call:

```
      if (use_extended_sk && primary_key < MAX_KEY && key &&
          !(keyinfo->flags & HA_NOSAME))
        key_part +=
            add_pk_parts_to_sk(keyinfo, key, share->key_info, primary_key,
                               share, handler_file, &usable_parts);
```

which updates part_of_key and part_of_sortkey as needed for PK fields so the code above isn't actually needed. Looking at the history for the if statement (with HA_PRIMARY_KEY_IN_READ_INDEX) it seems that it was done way back in 2005, so it is possible this is dead code (at least for storage engine that uses extended keys). I also verified a few extended key scenarios with Manuel and they work expected, further confirming this is the case.

Anyway, for now I've ported this patch but there is no way to validate it given that we don't have partition yet. We should validate this again once partition starts working.

Reviewed By: lth

Differential Revision: D19844310

fbshipit-source-id: 08e75a7f6e2