Remove skip unique check on lag and whitelist
Summary:
The skip unique check on lag feature isn't a particular useful feature and we don't use it on production. When porting MyRocks 8.0 we ported the myrocks portion which should be dropped. Note that we didn't port mysql unique_check_lag_* feature itself so this just removes the test and rocksdb portion.

Squash with D62421

Reviewed By: lth

Differential Revision: D19851812

fbshipit-source-id: 838b26653b1