Remove skip unique check on lag and whitelist
Summary: The skip unique check on lag feature isn't a particular useful feature and we don't use it on production. When porting MyRocks 8.0 we ported the myrocks portion which should be dropped. Note that we didn't port mysql unique_check_lag_* feature itself so this just removes the test and rocksdb portion.

Reviewed By: lloyd

Differential Revision: D19851812 (https://github.com/facebook/mysql-5.6/commit/44bb22216dcd072a70edfb2e33407e8f27aaab60)

fbshipit-source-id: 55303cccdc2