Make mysql_real_connect_nonblocking truly non-blocking
Summary:
The mysql_real_connect_nonblocking api is not non-blocking as it indicates.  To make it non-blocking the loop needs to only continue if the status is STATE_MACHINE_CONTINUE.  If the status is STATE_MACHINE_WOULD_BLOCK the function should return NET_ASYNC_NOT_READY.  The check for compression in the client flag should only occur the first time as all parameters (other than the mysql handle itself) are ignored on subsequent calls.  In addition the mysqltest was accessing the mysql->net field which might be null so we need to add a check for it.

Reported to Oracle: https://bugs.mysql.com/bug.php?id=98574

Reviewed By: jrahman

Differential Revision: D19864352

fbshipit-source-id: 929c5740506