Use upstream methods of restarting mysqld with options
Summary:
Some tests are failing because the wait for server shutdown is timing out in valgrind. We could extend the timeout, but I noticed that the upstream scripts automatically adjust the timeout if using valgrind.

Instead of using `include/restart_mysqld_with_option.inc` to restart mysqld, we should use the upstream version `include/restart_mysqld.inc` which automatically adjusts the timeout for valgrind.

Also lengthen the timeout for rocksdb.drop_table2.

Reviewed By: luqun

Differential Revision: D19949862

fbshipit-source-id: 78d8cff7e3b