Backport bugfix #29717909 from mysql-server/8.0
Summary:
Bug #29717909 MEMORY LIFETIME OF VARIABLES BETWEEN CHECK AND UPDATE INCORRECTLY MANAGED

PROBLEM

1) When setting innodb system variables of type string ,we use predefined callback
   functions check() and update() for checking the variable value and updating
   respectively.
2) For some variables in the check() part we were reading the input variable in a
   local buffer and then passing the pointer to this local buffer to the update() function
3) This caused an invalid read error in the valgrind since the address to local pointer
   is already lost when check() function ends.

FIX

1) After analyzing all the innodb system variables we found that these variables needed fixing
namely
   1. innodb_ft_server_stopword_table
   2. innodb_buffer_pool_filename
   3. innodb_ft_aux_table
   4. clone_valid_donor_list (for 8.0 version)
   5. innodb_redo_log_archive_dirs (for 8.0 version)
2) The fix is to allocate memory for this input variables from the thd->mem_root rather
than from the local buffer. Since all this variables are of the type PLUGIN_VAR_MEMALLOC,
server will copy these variables values to a separate buffer (managed by server) and free
the thd->mem_root at the end of statement.
3) We eliminated the update function for "innodb_ft_aux_table" by making this
variable as PLUGIN_VAR_MEMALLOC.
4) We eliminated the fts_internal_tbl_name2 variable which was used as intermediate variable
for updating the fts_internal_tbl_name. This was done to fix the BUG 24009272 SEGFAULT
WITH CREATE+SELECT FROM IS+DROP FTS TABLE CONCURRENTLY. The integrity of the variable is ensured by locking all the global variables before checking this variable.

RB: 23776, 23830
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>

Post push fix

1. innodb_buffer_pool_filename variable cannot be set to null in Windows. The behavior
is not consistent with Linux where there is no check function and it allows it be set
as NULL. It is better not to allow it to be set as NULL in Linux as well for the following reasons
   1. The purpose of the  innodb_buffer_pool_filename is to specify the name of the file
   where the list of tablespace IDs and page IDs produced by
   innodb_buffer_pool_dump_at_shutdown
   or innodb_buffer_pool_dump_now should go.
   2. When set as NULL in Linux and when we turn on  innodb_buffer_pool_dump_now
   it creates a file called "(null)" which is not correct.
So as a fix we will not allow innodb_buffer_pool_filename to be set as NULL
2. Doxygen failure in declaring the function name

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>

Squash with null

Reviewed By: hermanlee

Differential Revision: D21672580

fbshipit-source-id: b939c1e64f2