Fix rpl_mts.rpl_mts_stop_slave_report_pos
Summary:
This change (https://github.com/facebook/mysql-5.6/commit/64886ac) changed the error reporting in sql_slave_killed when transactions can be rolled back, and thus many previously more serious errors are not printed out to the error log.

This causes the pattern match for assert_only_after to fail, and we end up looking at logs from previous executions of this test (in different combinations), causing the match count to be higher than expected. The fix is to update assert_only_after with other reference points that we can use in the error log.

Squash with: D14387911

Reviewed By: abhinav04sharma, zhichengzhu

Differential Revision: D20040481

fbshipit-source-id: 42f020da215