Add more statistics(query_sample_text and query_sample_seen) to information_schema.sql_statistics
Summary:
In 8.0, we have tables in performance schema containing statistics for every normalized query: https://dev.mysql.com/doc/refman/8.0/en/statement-summary-tables.html

In 5.6, the analogue to that is information_schema.sql_statistics. However, we only support a subset of the stats. This task is to back port some of those stats into sql_statistics.

**What is new in this diff?**

1) Added support for fields `query_sample_text` and `query_sample_seen`.
2) Added a global variable named `max_digest_sample_age` which controls min re-sampling duration in seconds. A value equal to 0 means no re-sampling at all. Any other positive value, say 30, means to re-sample query is the last sampled query was run more than 30 seconds ago. Default value is set to 60 seconds. If set to -1, there is no query sampling.

Reviewed By: mzait

Differential Revision: D20099773

fbshipit-source-id: e2d61953639