Add resultset checksum feature
Summary:
Add feature to enable checksumming resultsets in a response.
Useful for validating proxies or other middleware that forwards
resultsets on behalf of clients. In this implementation, the row
contents themselves are checksummed. The ok packets and eof packets are
not, as it's a reasonable expectation the proxies or middleware may
modify those during normal operation.

Reviewed By: jkedgar

Differential Revision: D20199758

fbshipit-source-id: c815aa62862