Implement resultset checksums
Summary:
Implement a CRC32 checksum over the resultset field headers and
row contents. Enablement is controlled by a global variable + the query checksum feature being enabled.

The query_checksum attribute was renamed to checksum for symmetry. The idea being the an attribute with key 'checksum' contains a value reflecting a checksum of the entity (query or resultset) to which that attribute is attached.

Reviewed By: jkedgar

Differential Revision: D20186509 (https://github.com/facebook/mysql-5.6/commit/499a613f3610bcb9124ee5df74f5ad34e00ccd48)

fbshipit-source-id: 21b91a7736c