Ensure errors are propagated when exiting loop in QUICK_GROUP_MIN_MAX_SELECT::get_next
Summary:
In QUICK_GROUP_MIN_MAX_SELECT::get_next, when it tries to early exit the loop either when calculating min/max, it didn't set result to the failure. As a result the caller would end up thinking it returning a valid row, which resulting invalid values in record being returned - the record would only have the prefix key filled in this case, and the rest would be uninitialized / up to implementation of storage engine in the failure case. In MyRocks case, we always unpack the key into record in Rdb_key_def::pack_index_tuple, so even in the case that when the key is not found we would still unpack the key into record and end up returning that to the caller as valid row. In InnoDB this issue doesn't repro because InnoDB doesn't eagerly unpack key into record. You could argue that MyRocks should not pollute the record in the failure case, ideally. Regardless, it is incorrect for the MySQL layer to depend on the value of record in the failure case.

In addition to setting result in early exit, if the last iteration happens to early exit, we also need to reset result to 0 if we have found other results earlier. There is already code doing that but we need to move it outside of the loop. Note that this means that we would still be returning an invalid record from this failure, which shall be addressed in a future change, most likely by saving previously success record contents.

This seems to be a regression from 5.6 introduced by WL#13066 Improve predicate handling in loose index scan for group by, which we back ported from 8.0.19 to 8.0.17.

With this fix the tests results are showing we are aligning with 5.6 behavior.

Reviewed By: lth

Differential Revision: D20187102

fbshipit-source-id: 1c2ff3c1a37