Stabilize rocksdb.drop_cf_during_show_global_info in valgrind
Summary:
The test looks to have a timing issue where a table and column family
are being dropped, but the test had not detected the drop is still
ongoing, causing the create table statement for that cf to fail.

Fix this by checking to see if the drop_cf list is empty before allowing
the test to continue.

Reviewed By: lloyd

Differential Revision: D20290539