Fix client build for TP2 rebase
Summary:
Running a build with -DWITHOUT_SERVER=1 and -DDISABLE_SHARED failed because the CMake settings attempted to get information from a shared library that was not built (because DISABLE_SHARED was specified.  This change resolves the problem.

No bug report was generated for Oracle because the DISABLE_SHARED option was removed from MySQL in 8.0.18.

Reviewed By: jrahman

Differential Revision: D20403075

fbshipit-source-id: 302762d67b9