WL#13929: Introduce new optimizer switch to disable limit optimization Bug#30348211: DANGEROUS OPTIMIZATION RECONSIDERING_ACCESS_PATHS_FOR_INDEX_ORDERING
Summary:
The optimization to switch from a non-ordering index to an
ordering index for group by and order by when there is a
limit clause makes wrong decisions for some queries for various
reasons.
This patch introduces an optimizer switch that enables users
to disable this optimization if the switch to ordering index
results in a bad plan.
The name of the switch is "prefer_ordering_index". By default the
switch is on which means if an ordering index is found to be better
than an already chosen plan then optimizer will switch to the ordering
index. However if "prefer_ordering_index" is set to off, the switch
to ordering index because of the "limit optimization" will not happen,
resulting in optimizer picking the original plan which could have been
using a filtering index.

Approved By : Sergey Glukhov <sergey.glukhov@oracle.com>

Notes:
Backported from https://github.com/mysql/mysql-server/commit/ce91d514bd1e92c6136a9187826f10b0c33463e3

This provides the equivalent of the `optimizer_low_limit_heuristic` variable in 5.6.

Reviewed By: luqun

Differential Revision: D27893587

fbshipit-source-id: 1f26a64e737