Support older compression libraries
Summary: For open source users, they may not have the latest compression library versions installed on their machines. Since we're not using any new functionality in the new releases and only using some convenience functions to save on allocations, we can just deallocate and reallocate the compression contexts as needed. There is on additional change where the lz4 header size was not available in prior releases. The exact number isn't critical for correctness, so I just used 19 (the actual value).

Reviewed By: terrelln

Differential Revision: D20748160

fbshipit-source-id: 2d10b614f31