Fix main.mysql_80_inplace_upgrade
Summary:
In a previous change (D20454576), we added code to backup and restore certain tables in the mysql schema. However the restore code would not run if the change bailed out at any point during the test (eg. if we want to skip the test due to valgrind). This meant that we were leaving temporary backed up tables around, breaking mtr's internal check.

Fix this by moving the backup code to happen after we verify that the test is eligible to be run.

Squash with: D20195466

Reviewed By: hermanlee

Differential Revision: D20819467

fbshipit-source-id: 19a6a4ba155