Rotate binlogs when setting gtid_purged
Summary:
In 5.6, the binlog was rotated when gtid_purge was set, but this
behavior changed in 8.0. Functionality such as SHOW GTID_EXECUTED IN
depend on reading the prev_gtid event of the binlog to construct the
executed set up to that point in the binlog file. If gtid_purge was set
without rotating the binlog, SHOW GTID_EXECUTED IN does not include the
purged gtid information.

Squash with D9368239

Reviewed By: yizhang82

Differential Revision: D20842600

fbshipit-source-id: b91da8cb25b