Fix main.ssl-big
Summary:
This test tries to create connections in a loop in order to reproduce a bug, but does not wait until connections are actually disconnected before moving on to the next loop. Since the default max connection limit is 151 and we're looping 2000 times, it's possible for us to run out of connection slots on busy servers.

Fix by bumping the max connection limit to 2500.

Reviewed By: yizhang82

Differential Revision: D20842347

fbshipit-source-id: d287822482c