Backport Bug#30088404 : CONCURRENT LOG BUFFER OF 8.0 CAUSES REGRESSION OF CPU BOUND PERFORMANCE ON LOW-END ARCHITECTURE
Summary:
This backports the InnoDB feature to reduce redo logging cpu.
innodb_log_writer_threads enables/disables the use of the new writer log
threads. When disabled, each user thread performs redo log flushing.

Porting notes

Feature is in 8.0.22, so can be dropped when rebasing to 8.0.22
Squash with null when rebasing to 8.0.22

Differential Revision: D23301452

fbshipit-source-id: 69d2a0b447e