Enabled the --upgrade=FORCE_AND_SHUTDOWN option in the upgrade path
Summary:
Enabled the --upgrade=FORCE_AND_SHUTDOWN option in the upgrade path.

**Background**
This is added to support the upgrades in 8.0 mysql. With 8.0 mysql, mysql_upgrade tool is deprecated and all the options are appended in mysqld for upgrades. Current dba upgrade script calls mysql_upgrade and then start mysql with mysqld_safe script.
To fix upgrades in 8.0, there are 2 options -
1) mysqld_safe start mysql with --upgrade=FORCE on upgrades. This is not trivial because mysqld_safe is called by linux systemd which allows us to pass only single parameter(port). Passing additional parameter is not trivial as it requires the usage sticky env file which poses additional threats to the regular restart use cases.
2) Another option was to replace the mysql_upgrade script with mysqld invocation which upgrades the server and then safely exits. Mysql start will be taken care by mysqld_safe script.

After looking at the pro and cons, we have picked option2.

Differential Revision: D20939322

fbshipit-source-id: d8cb02e9bd4