Do a reset master after setting up the initial raft ring
Summary:
When you enable raft and the first leader election happens, the instances' binlog files are rotated. Tests which do "SHOW BINLOG EVENTS" and rely on testing binlog events in the particular binlog files will fail. For example, if a test does some writes to a table, these events will be found in the first binlog file in non raft mode whereas in the second binlog file in raft mode.

We should do a `reset master` after the ring is set up so that the instances go back to binlog 00001

Reviewed By: abhinav04sharma

Differential Revision: D20954461

fbshipit-source-id: 1b3e50b4641