Bug#31168097: SIG 6 ASSERTION `0' IN TIME_ZONE_UTC::GET_NAME AT TZTIME.CC:1182
Summary:
Filesort needs to sort temporal values based on their UTC
equivalents, so it temporarily alters the session's time
zone. If another session is listing the altered session
variable from performance_schema, it will see UTC. Now,
internally we have a special class for the UTC time zone
(don't ask me why) that has assertions in place for most
operations that it isn't intended for, it's really only
intented for internal conversion to and from UTC. One of
those assertions now fires. But this is a blessing in
disguise, since the optimizer is not supposed to tamper with
the session's time zone, or any session variable, for that
matter.

This fix borrows an idea originally from wl#12535 that adds
a new interface for getting times in UTC. The difference is
only the use of it. There, the rationale was to present
TIMESTAMP values the way they were stored - in UTC - to the
user. In this case the audience is filesort, which wants the
packed values in UTC, so it can sort on the same thing that
indexes or comparators see when they sort. But the low-level
change is the same: We add an interface to
Field_temporal_with_date that brings out the UTC value
without conversion. Then we add two new Item interface
member functions that use those. We also move the function
get_int_sort_key_for_item_inline() to Item.

There's no test case, as it would have been rather contrived
for this bug, since it's a rare race condition and now there
is no race anymore.

As a bonus, we make Time_zone::gmt_sec_to_TIME() const, the
author obviously forgot that.

This was backported from 8.0.22 and can be dropped when rebasing to
8.0.23 or later.

Squash with null

Reviewed By: Pushapgl

Differential Revision: D32401633

fbshipit-source-id: ee87eb9abab