avoid extra file rotation during startup
Summary:
A mysqld server, on startup finds that the latest  binlog is marked as 'closed'
and hence a new binlog was opened. The file was marked as closed because during
promotion, the no-op rotate event is written to the relay log and the rotation
mechanism works through relay log flow. Relay logs are seq append caches and
are always created with the header marked as 'closed'. This code changes this
behavior where relay logs are always marked as 'opened' in raft mode rotation.
From mysql's standpoint, this seems perfetcly fine since that bit is almost
useless for the relay log mechanism today.

Reviewed By: anirbanr-fb

Differential Revision: D21180421

fbshipit-source-id: 425197b09c6