Subtract 1 from binlog name if raft is enabled
Summary:
Due to an unavoidable binlog rotation that happens when the first leader is elected, binlog files in all instances are moved ahead by 1. This makes all tests that do "SHOW BINLOG EVENTS" fail because in raft mode, the binlog name that appears in the result is one "ahead" than what is expected.

To fix this, we could do "RESET MASTER" on all instances. However, that causes more problems for the OPID index in MyRaft.

For now, if raft is enabled, we can hack into the output of "SHOW BINLOG EVENTS" to subtract 1 from the binlog file name. This will make it appear as if no rotation happened in the beginning.

Reviewed By: abhinav04sharma

Differential Revision: D21194040

fbshipit-source-id: 72d60794fbc